requirements.txt
flask
aspose-words
pypandoc